controller for poke-express
